requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
github3.py==1.3.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jiter==0.7.1
jwcrypto==1.5.6
openai==1.54.4
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGithub==2.5.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
unidiff==0.7.5
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.16.0
